protecting fetch calls and clearing unnecessary code

+++ src/app/api/admin/route.js
@@ -1,7 +1,21 @@
 import { prisma } from "@/lib/prisma";
+import { cookies } from "next/headers";
 import { NextResponse } from "next/server";
 
 export async function GET(request) {
+  const cookieStore = cookies();
+  const userCookie = cookieStore.get("user");
+  if (!userCookie) {
+    return new Response("Unauthorized ", {
+      status: 401,
+    });
+  }
+  const currentUser = JSON.parse(userCookie.value);
+  if (!currentUser.role === "ADMIN") {
+    return new Response("Unauthorized user", {
+      status: 401,
+    });
+  }
   const patients = await prisma.patient.findMany({
     include: {
       treatments: true,
@@ -12,6 +26,19 @@ export async function GET(request) {
 }
 
 export async function PUT(request) {
+  const cookieStore = cookies();
+  const userCookie = cookieStore.get("user");
+  if (!userCookie) {
+    return new Response("Unauthorized ", {
+      status: 401,
+    });
+  }
+  const currentUser = JSON.parse(userCookie.value);
+  if (!currentUser.role === "ADMIN") {
+    return new Response("Unauthorized user", {
+      status: 401,
+    });
+  }
   const data = await request.json();
   await prisma.patientTreatment.deleteMany({
     where: {
@@ -37,8 +64,20 @@ export async function PUT(request) {
 }
 
 export async function DELETE(request) {
+  const cookieStore = cookies();
+  const userCookie = cookieStore.get("user");
+  if (!userCookie) {
+    return new Response("Unauthorized ", {
+      status: 401,
+    });
+  }
+  const currentUser = JSON.parse(userCookie.value);
+  if (!currentUser.role === "ADMIN") {
+    return new Response("Unauthorized user", {
+      status: 401,
+    });
+  }
   const patientId = await request.json();
-  console.log(patientId);
   await prisma.patient.delete({
     where: {
       id: patientId,
